+++ interpreter/src/memory.ts
@@ -1,22 +1,28 @@
 import { EventType, EventsInstance } from './events';
 
 const SPECIAL_VARIABLE_NAMES = [ 'OPTIONS_COUNT' ];
+const EXTERNAL_VARIABLE_PREFIX = '@';
 
 export type InternalMemory = {
   access: any;
   variables: any;
   internal: any;
+  e_variables: any;
 }
+
+export type DialogueData = Omit<InternalMemory, "e_variables">;
 
 export interface MemoryManager {
     setAsAccessed(id: string): void;
     wasAlreadyAccessed(id: string): boolean;
     getVariable<T>(id: string, defaultValue?: T): T;
     setVariable<T>(id: string, value: T): T;
+    getExternalVariable<T>(id: string, defaultValue?: T): T;
+    setExternalVariable<T>(id: string, value: T): T;
     setInternalVariable<T>(id: string, value: T): T;
     getInternalVariable<T>(id: string, defaultValue?: T): T;
-    getAll(): InternalMemory;
-    load(data: InternalMemory): void;
+    getAll(): DialogueData;
+    load(data: DialogueData): void;
     clear(): void;
 }
 
@@ -24,7 +30,8 @@ export function Memory(listeners: EventsInstance, init?: InternalMemory): Memory
   let mem: InternalMemory = init || {
     access: {},
     variables: {},
-    internal: {}
+    internal: {},
+    e_variables: {},
   };
 
   return {
@@ -41,6 +48,10 @@ export function Memory(listeners: EventsInstance, init?: InternalMemory): Memory
         return this.getInternalVariable(id, defaultValue);
       }
 
+      if (id.startsWith(EXTERNAL_VARIABLE_PREFIX)) {
+        return this.getExternalVariable(id, defaultValue);
+      }
+
       const value = mem.variables[id];
       if (value === undefined) {
         return defaultValue;
@@ -49,8 +60,27 @@ export function Memory(listeners: EventsInstance, init?: InternalMemory): Memory
     },
 
     setVariable<T>(id: string, value: T): T {
+      if (id.startsWith(EXTERNAL_VARIABLE_PREFIX)) {
+        return this.setExternalVariable(id, value);
+      }
+
       listeners.triggerEvent(EventType.VARIABLE_CHANGED, { name: id, value, previousValue: mem.variables[id] });
       return mem.variables[id] = value;
+    },
+
+    getExternalVariable(id: string, defaultValue?: any): any {
+      const sanitizedId = id.replace(EXTERNAL_VARIABLE_PREFIX, "");
+      const value = mem.e_variables[sanitizedId];
+      if (value === undefined) {
+        return defaultValue;
+      }
+      return value;
+    },
+
+    setExternalVariable<T>(id: string, value: T): T {
+      const sanitizedId = id.replace(EXTERNAL_VARIABLE_PREFIX, "");
+      listeners.triggerEvent(EventType.EXTERNAL_VARIABLE_CHANGED, { name: sanitizedId, value, previousValue: mem.e_variables[sanitizedId] });
+      return mem.e_variables[sanitizedId] = value;
     },
 
     setInternalVariable<T>(id: string, value: T): T {
@@ -65,19 +95,27 @@ export function Memory(listeners: EventsInstance, init?: InternalMemory): Memory
       return value;
     },
 
-    getAll(): InternalMemory {
-      return mem;
+    getAll(): DialogueData {
+      return {
+        variables: mem.variables,
+        access: mem.access,
+        internal: mem.internal,
+      };
     },
 
-    load(data: InternalMemory): void {
-      mem = data;
+    load(data: DialogueData): void {
+      mem = {
+        ...data,
+        e_variables: {},
+      };
     },
 
     clear(): void {
       mem = {
         access: {},
         variables: {},
-        internal: {}
+        internal: {},
+        e_variables: {},
       };
     },
   };